+++ lib/src/util/constants.dart
@@ -1,0 +1,2 @@
+const setUpMethodName = 'bddSetUp';
+const tearDownMethodName = 'bddTearDown';