+++ test/step_folder_name_test.dart
@@ -20,8 +20,8 @@ import 'package:flutter_test/flutter_test.dart';
 import './../../../custom_steps/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await theAppIsRunning(tester);
     });
   });