+++ test/tag_test_method_name_test.dart
@@ -21,8 +21,8 @@ import 'package:flutter_test/flutter_test.dart';
 import './step/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    customTestWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    customTestWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await theAppIsRunning(tester);
     });
   });
@@ -60,11 +60,11 @@ import 'package:flutter_test/flutter_test.dart';
 import './step/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    customTestWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    customTestWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await theAppIsRunning(tester);
     });
-    otherTestWidgets('Testing scenario 2', (tester) async {
+    otherTestWidgets(\'\'\'Testing scenario 2\'\'\', (tester) async {
       await theAppIsRunning(tester);
     });
   });