+++ test/test_method_name_test.dart
@@ -20,8 +20,8 @@ import 'package:flutter_test/flutter_test.dart';
 import './step/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    customTestWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    customTestWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await theAppIsRunning(tester);
     });
   });